+++ apps/curator-api/src/curator/utils.ts
@@ -1,0 +1,15 @@
+const resolveTwitterShortUrl = async (url?: string) => {
+  try {
+    console.log("resolveTwitterShortUrl", url);
+    if (url == null) return null;
+
+    const resolved = await fetch(url, { method: "HEAD", redirect: "follow" });
+
+    return resolved.url;
+  } catch (e) {
+    console.error("Error resolving Twitter short URL:", e);
+    return null;
+  }
+};
+
+export { resolveTwitterShortUrl };