+++ apps/curator-api/src/express.d.ts
@@ -1,0 +1,20 @@
+import type { Profile } from "passport-twitter";
+import "express";
+
+// Tell Express that User is a Profile from passport (e.g. contains Twitter username etc.)
+declare global {
+  namespace Express {
+    interface User extends Profile {
+      name: string;
+      id_str: string;
+      _json: {
+        description: string;
+        profile_image_url_https: string;
+        profile_banner_url: string;
+        url: string;
+      };
+    }
+  }
+}
+
+export {};