+++ packages/env/src/utils.ts
@@ -3,7 +3,7 @@ import { cloakServers, Environment, ipfsServers } from ".";
 const getCloakServer = (env?: Environment) =>
   cloakServers[env ?? Environment.PRODUCTION];
 
-const getMetadataServer = (env?: Environment) =>
+const getCuratorServer = (env?: Environment) =>
   ipfsServers[env ?? Environment.PRODUCTION];
 
-export { getCloakServer, getMetadataServer };
+export { getCloakServer, getCuratorServer };