+++ packages/sdk/src/core/create/types.ts
@@ -6,6 +6,8 @@ import { AuctionMetadataSchema } from "./schema";
 
 type CreateParams = v.InferInput<typeof schema>;
 type CreateConfig = ContractConfig<typeof abis.batchAuctionHouse, "auction">;
-type AuctionMetadata = v.InferInput<typeof AuctionMetadataSchema>;
+type AuctionMetadata = v.InferInput<typeof AuctionMetadataSchema> & {
+  toString: () => string;
+};
 
 export type { CreateParams, CreateConfig, AuctionMetadata };